backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
openai==1.3.5
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
watchdog==3.0.0
requests==2.31.0
google-generativeai==0.3.2